Noise correction API

CaAnalysis.process_traces — Function

process_traces(
    param::Dict, activity_traces::Dict, marker_traces::Dict, threshold::Real, t_range; 
    activity_bkg=nothing, marker_bkg=nothing, min_intensity::Real=0, interpolate::Bool=false, 
    denoise::Bool=false, bleach_corr::Bool=false, divide::Bool=false, normalize_fn::Function=x->quantile(x,0.2),
    k::Union{Real,Nothing}=nothing, valid_rois=nothing
)

Applies multiple data processing steps to the traces. The order of processing steps is:

Background-subtraction

Delete low S/N neurons

Delete neurons detected in too few time points

Interpolate traces at missing data points

Denoise

Bleach-correct

Divide activity by marker channel

Normalize

Deconvolve

Zscore

Arguments

param::Dict: Parameter dictionary containing laser information.

activity_traces::Dict: traces in the activity channel

marker_traces::Dict: traces in the marker channel

threshold::Real: Number of timepoint detections necessary to include a neuron in the analysis

t_range: Time points to extract traces over.

Keyword arguments

activity_bkg: Background in activity channel. If left blank, background will not be subtracted.

marker_bkg: Background in marker channel. If left blank, background will not be subtracted.

min_intensity::Real: Minimum average intensity in the activity channel for a neuron (after background subtraction). Neurons with less than this much signal will be removed. Default 0.

interpolate::Bool: Whether to interpolate missing data points within the time range.

denoise::Bool: Whether to apply a total variation denoising step.

bleach_corr::Bool: Whether to bleach-correct the traces.

divide::Bool: Whether to divide the activity channel traces by the marker channel traces.

normalize_fn::Function: Function to use to get "average" activity when normalizing traces.

k::Union{Real,Nothing}: Deconvolution parameter. Set this to (time length of confocal volume) / (GCaMP decay half-life)

CaAnalysis.divide_by_marker_signal — Function

divide_by_marker_signal(traces::Dict, marker_traces::Dict)

Divides the activity traces traces::Dict by the marker signal marker_traces::Dict

CaAnalysis.get_background — Function

get_background(timepts, get_basename::Function, nrrd_path::String, ch::Int)

Gets the background camera intensity from the NRRD files to be subtracted.

Arguments

timepts: Time points to get background of

get_basename::Function: Function that outputs the name of the NRRD file from the time and channel

nrrd_path::String: Path to NRRD files

ch::Int: Channel

CaAnalysis.bkg_subtract — Function

bkg_subtract(traces::Dict, frame_bkg)

Subtracts from traces the corresponding background frame_bkg.

CaAnalysis.normalize_traces — Function

normalize_traces(traces; zero::Bool=false, fn::Function=mean)

Normalizes traces.

Arguments

traces::Dict: Traces to normalize

Optional keyword arguments

zero::Bool: Whether to normalize to zero (so activities that are less than average would be negative). Default false.

fn::Function: Function that determines average neuron activity

CaAnalysis.interpolate_traces — Function

interpolate_traces(traces::Dict, t_range; itp_method=Linear(), extrap_method=Interpolations.Flat(), fill_val=NaN)

Interpolates data at missing time points.

Arguments

traces::Dict: Dictionary containing for each neuron a set of time points and activity values at those time points

t_range: Time points to interpolate to. Time points outside this range will be set to fill_val

itp_method (optional, default Linear()): Interpolation method.

extrap_method (optional, default Interpolations.Flat()): Extrapolation method (for interpolating data points outside all time points the neuron was detected in)

fill_val (optional, default NaN): Value to fill time points outside the interpolation range.

CaAnalysis.zscore_traces — Function

zscore_traces(traces)

Z-scores traces.

CaAnalysis.get_all_values — Function

get_all_values(processed_traces)

Gets all values from a set of processed traces. Ensure that the traces have been normalized prior to using this function!

CaAnalysis.get_laser_intensity — Function

get_laser_intensity(percent_on, laser)

Interpolate laser intensity from percentage on, and from measurements.

get_laser_intensity(percent_on, intensity, laser_perc; zero_thresh=4.5, min_laser=5.2, max_laser=20.0, max_interpolate=2.0)

Interpolate laser intensity from percent_on, and from intensity measurments taken at laser_perc values.

Arguments:

percent_on: Laser percentage used in data acquisition.

intensity: Measured calibration intensities.

laser_perc: Measured laser percentages corresponding to those intensities.

zero_thresh (optional, default 4.5): Maximum percentage to use for 0 calibration

min_laser (optional, default 5.2): Minimum percentage to use for linear calibration

max_interpolate (optional, default 2.0): Maximum laser percentage difference to use for linear interpolation

CaAnalysis.fit_bleach — Function

fit_bleach(f, t, plot_fit=true, use_mono=false, quantile_norm=0.5)

Fits double exponential bleaching model.

Arguments

f: 1D data to fit the bleaching model

t: Time points

plot_fit: plot fit result if true

use_mono: use mono exponential model instead of double exponential model if true

fit_bleach(f::Array{<:Real,2}, plot_fit=true, use_mono=false; idx_t=:all)

Calculates mean activity across the units and fits the double exponential bleaching model.

Arguments

f: N x T data array. N: number of units, T: number of time points.

plot_fit: plot fit result if true

use_mono: use mono exponential model instead of double exponential model if true

CaAnalysis.fit_bleach! — Function

fit_bleach!(data_dict::Dict, plot_fit=true, use_mono=false; data_key="f_denoised", idx_unit=:ok, idx_t=:all)

Calculates mean activity across the units and fits the double exponential.

Arguments

data_dict: data_dictionary.

plot_fit: plot fit result if true

idx_unit: see get_idx_unit()

idx_t: see get_idx_t()

data_key: key of data_dict to be used for fitting the model

use_mono: use mono exponential model instead of double exponential model if true